payment moved to farmer
