+++ Project Work/index.js
@@ -24,7 +24,6 @@ app.use(cors());
 const admin = require('./routes/admin');
 const supervisor = require('./routes/super');
 const farmer = require('./routes/farmer');
-const payment = require('./routes/payment');
 
 const auth = require('./routes/auth');
 
@@ -37,7 +36,7 @@ app.use(`${version}/admin`, admin);
 app.use(`${version}/supervisor`, supervisor);
 app.use(`${version}/farmer`, farmer);
 app.use(`${version}/auth`, auth);
-app.use(`${version}/payment`, payment);
+
 
 app.get('/api/', (req, res) => res.send({ data: 'Meet Shah' }));
